ia_agent/requirements.txt
# API Web
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.7.0

# Client HTTP pour appeler les autres APIs
httpx==0.27.0

# Détection de langue
langdetect==1.0.9

# Database
psycopg2-binary==2.9.9
